refactor: 💡 almost done salseman screens
