+++ package.json
@@ -25,7 +25,6 @@
     "lint-staged": "^12.3.3",
     "prettier": "^2.5.1",
     "eslint": "^8.10.0"
-
   },
   "dependencies": {
     "eslint-plugin-unused-imports": "2.0.0"
