+++ packages/frontend/src/components/ui/salesman/membersDetails/MembersHeader/MembersHeader.style.js
@@ -1,0 +1,5 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({}));
+
+export default useStyles;
